Create context module
