+++ snippets/src/language.rs
@@ -7,3 +7,4 @@
 pub mod c99_tc3;
 
 mod normalize_comments;
+mod context;
